moving fetchNodeKeys to canisterStatus
